// src/modules/vacancies/components/VacancyFormRePost/VacancyFormRePost.types.ts
export interface VacancyFormRePostTypes {
  tags?: string[];
  remote?: string;
  description: string;
  requirements: string;
  responsibilities: string;
  termsAndConditions: string;
  name: string;
  other: string;
  qualification: string;
  experience: number;
  mode: string;
  salary: number;
}